refactor : 각 컴포넌트의 socket 비즈니스 로직을 커스텀훅으로 분리
- webSocket 비즈니스 로직을 따로 분리하여, 코드의 흐름을 파악하기 용이하도록 만듦
