+++ frontend/src/components/landing/memo/MemoBlock.tsx
@@ -1,7 +1,7 @@
 import MemoEditor from "./MemoEditor";
 import { LandingMemoDTO } from "../../../types/DTO/landingDTO";
 import { MemoColorStyle, MemoColorType } from "../../../types/common/landing";
-import useLandingMemo from "../../../hooks/common/landing/useLandingMemo";
+import useLandingMemo from "../../../hooks/common/landing/useLandingMemoEditor";
 import React from "react";
 
 interface MemoBlockProps extends LandingMemoDTO {
